pyproject.toml
[tool.poetry]
name = "addressbook"
version = "0.1.0"
description = ""
authors = ["Mirek Zalewski <mz@miroslaw-zalewski.eu>"]

[tool.poetry.dependencies]
python = "^3.8"
PyQt5 = "^5.15.1"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
addressbook = "addressbook.addressbook:main"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
